==> 5-bcell_analisis/19-analysing_bcell_level_1.Rmd <==
---
title: "19-Analysing clusters B cells/ LinkPeaks Genome Wide"
author: 
- name: "Maria Lucia Romero Rivero"
  email: MLROMEROR@clinic.cat
date: "`r Sys.Date()`"
output:
  BiocStyle::html_document:
    toc: true
    toc_float: true
    number_sections: true
    fig_captions: yes

---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Pre-processing 

## Load package


```{r message=FALSE, warning=FALSE}
library(RColorBrewer)
library(Signac)
library(Seurat)
library(GenomicRanges)
library(future)
library(harmony)
library(EnsDb.Hsapiens.v86)
library(stringr)
library(dplyr)
library(ggplot2)
library(ggpubr)
library(patchwork)
library(kableExtra)
library(tidyverse)
library(BSgenome.Hsapiens.UCSC.hg38)
library(plyr)
library(biomaRt)
library(BiocManager)
library(GenomicRanges)
library(ComplexHeatmap)
library(binovisualfields)
library(viridis)
set.seed(123)
```


## Parameters

```{r}
# Paths
path_to_obj <- ("~/Documents/multiome_tonsil_Lucia/results/R_objects/18.tonsil_multiome_bcell_level_1.rds")
path_to_markers<-("~/Documents/multiome_tonsil_Lucia/results/tables/18_tonsil_markers_level_1.csv")
path_to_obj_save <- ("~/Documents/multiome_tonsil_Lucia/results/R_objects/19.tonsil_multiome_bcell_level_1_Genwide_Linkpeak.rds")
path_to_save_DARs<-("~/Documents/multiome_tonsil_Lucia/results/tables/19.df_DARs_level_1.csv")
path_to_save_join_link_DARs <- ("~/Documents/multiome_tonsil_Lucia/results/tables/19.df_joint_link_DARs_level_1_Genwide.csv")
path_to_save_join_link_DARsDEA<-  ("~/Documents/multiome_tonsil_Lucia/results/tables/19.df_joint_link_DARs_DEA_level_1_Genwide.csv")
path_to_obj_save_dars_dea <- ("~/Documents/multiome_tonsil_Lucia/results/R_objects/19.tonsil_multiome_bcell_level_1_Genwide_Linkpeak_DARs_DEA.rds")
path_to_save_markers<-("~/Documents/multiome_tonsil_Lucia/results/tables/18_tonsil_markers_level_1_namecluster.csv")

```


## Load data

```{r}
tonsil_wnn_bcell <- readRDS(path_to_obj)

tonsil_markers<-read_csv(path_to_markers)
```

# UMAP 

```{r}
 DimPlot(
    tonsil_wnn_bcell,
    group.by = "wsnn_res.0.05",
    reduction = "wnn.umap",
    pt.size = 0.1, label = T
  )
```


```{r}
 DimPlot(
    tonsil_wnn_bcell,
    group.by = "Phase",
    reduction = "wnn.umap",
    pt.size = 0.1, label = T
  )
```

# Rename

```{r}
Idents(tonsil_wnn_bcell)<-"wsnn_res.0.05"
new.cluster.ids <- c("NBC","MBC", "GCBC", "GCBC","PC")

names(new.cluster.ids) <- levels(tonsil_wnn_bcell)
tonsil_wnn_bcell <- RenameIdents(tonsil_wnn_bcell, new.cluster.ids)

new.level<-c("NBC","GCBC","MBC", "PC")
tonsil_wnn_bcell@active.ident <- factor(tonsil_wnn_bcell@active.ident, levels = new.level)

cluster_color<-setNames(c("#A6CEE3","#B2DF8A", "#1F78B4", "#33A02C"),
                        levels(tonsil_wnn_bcell))

tonsil_wnn_bcell[["annotation_level_2"]] <- Idents(object = tonsil_wnn_bcell)

DimPlot(tonsil_wnn_bcell,label = F, reduction = "wnn.umap", pt.size = 0.5,cols = cluster_color) + ggtitle("Joint UMAP")
```
### Save to PDF

```{r}
library(raster)
p <- DimPlot(tonsil_wnn_bcell_links_DARs_DEA,label = F, reduction = "wnn.umap", pt.size = 0.5,cols = c("#a6cee3", "#1f78b4","#b2df8a", 
             "#33a02c", "#fb9a99","#e31a1c")) + ggtitle("Joint UMAP")
#rasterize(p, dpi = 300)
pdf(paste("/Users/mlromeror/Documents/multiome_tonsil_Lucia/Figures/
", "UMAP_bcell_data_res_0.05_legend.pdf", sep=""),width=5,height=5,paper="special")
p
dev.off()
```

## Number of cell per cell type

```{r}
cell<-as.data.frame(table(tonsil_wnn_bcell$annotation_level_2))
cell
```

```{r}
barp<-ggbarplot(cell,"Var1","Freq",fill  = "Var1",color = "Var1",palette = cluster_color,label = TRUE, lab.pos = "out", lab.col = "black",xlab = FALSE,ylab = "number of cells")
pdf(paste(figures_folder, "barplot_numbercell.pdf", sep=""),width=5,height=5,paper='special')


ggpar(barp,
      legend.title = "cell type")
dev.off()
```
```{r}
ggpar(barp,
      legend.title = "cell type")

```
## Get top n markers of each cluster


**Level 1=  Resolution 0.05**
 
```{r}

newValueVector <- c("0" = "NBC", "1" = "MBC","2"="GC","3"= "GC","4"="PC")
tonsil_markers$cluster <- as.character(tonsil_markers$cluster)
tonsil_markers$cluster <-  revalue(tonsil_markers$cluster, newValueVector)
write_csv(tonsil_markers,path_to_save_markers)

top5_tonsil_markers<-tonsil_markers %>% group_by(cluster) %>% top_n(n = 5, wt = avg_log2FC)
top7_tonsil_markers<-tonsil_markers %>% group_by(cluster) %>% top_n(n = 7, wt = avg_log2FC)
top10_tonsil_markers<-tonsil_markers %>% group_by(cluster) %>% top_n(n = 10, wt = avg_log2FC)
```

```{r}
kbl(top7_tonsil_markers,caption = "Table of the top 7 marker of each cluster resolution 0.005") %>%
  kable_paper("striped", full_width = F)
```

### Dynamic table top 7

```{r}

DT::datatable(top7_tonsil_markers)

```


### Dynamic table of all markers

```{r}
DT::datatable(tonsil_markers)

```


### Dotplot

```{r}
dot.10 <- DotPlot(tonsil_wnn_bcell, features = unique(top10_tonsil_markers$gene),cols = c("lightgrey","blue"), cluster.idents = T) + theme(axis.text.x = element_text( size = 10, vjust = 0.8, hjust = 0.8)) +ggtitle("Level 1 top 10 markersof each cluster")

dot.5 <- DotPlot(tonsil_wnn_bcell, features = unique( top5_tonsil_markers$gene),cols = c("lightgrey","blue"), cluster.idents = T) + theme(axis.text.x = element_text( size = 10, vjust = 0.8, hjust = 0.8)) + ggtitle("Level 1 top 5 markers of each cluster")

```

```{r}

dot.10 +
  coord_flip() +
  theme(axis.title = element_blank(), axis.text.y = element_text(size = 5))
dot.5 +
  coord_flip() +
  theme(axis.title = element_blank(), axis.text.y = element_text(size = 7))

```



#### Dotplot of Known markers

```{r}
markers<-c("BANK1", "FCER2","CD3D", "IL7R","MKI67", "TOP2A","MARCKSL1", "RGS13", "LMO2", "CCDC88A","GNLY", "NKG7", "GZMK", "CD8A",	"FCRL4", "FCRL5", "PLAC8", "SOX5","IGHG1", "IGHA1", "JCHAIN", "XBP1","LYZ", "S100A8","MT2A", "CD3D", "TRAC", "PCNA","CD19","CR2","MS4A1","RALGPS2","CD79A")


DefaultAssay(tonsil_wnn_bcell)<-"RNA"

dot <- DotPlot(tonsil_wnn_bcell, features = unique(markers),cols = c("lightgrey", "blue"), cluster.idents = T) + theme(axis.text.x = element_text( size = 10, vjust = 0.8, hjust = 0.8)) + ggtitle("Known markers")

dot +
  coord_flip() +
  theme(axis.title = element_blank(), axis.text.y = element_text(size = 5))


```
```{r}
m<-c("SUGCT", "MEF2B", "BCL6", "AICDA","TCL1A", "SOX5","BCL2A1","CD27", "MZB1","XBP1","IRF4","PRDM1","SLAMF7", "MZB1", "FKBP11", "FCRL5", "IGHM","IGHD","FCER2","CXCR4","BANK1","CD83", "PLAC8","FCRL4")
DefaultAssay(tonsil_wnn_bcell_links_DARs_DEA)<-"RNA"

dot <- DotPlot(tonsil_wnn_bcell, features = unique(m),cols = c("lightgrey", "blue"), cluster.idents = T) + theme(axis.text.x = element_text( size = 10, vjust = 0.8, hjust = 0.8)) + ggtitle("Known markers")

dot +
  coord_flip() +
  theme(axis.title = element_blank(), axis.text.y = element_text(size = 5))
```



# Scaling data

```{r message=FALSE, warning=FALSE}
all.genes <- rownames(tonsil_wnn_bcell)
tonsil_wnn_bcell <- ScaleData(tonsil_wnn_bcell, features = all.genes)
```

```{r}
DoHeatmap(tonsil_wnn_bcell,
          features=m,
           angle = 0,label = FALSE)
```

```{r}
sum(markers %in% rownames(GetAssayData(tonsil_wnn_bcell , slot = 'scale.data')))

```


```{r}
DoHeatmap(tonsil_wnn_bcell, features = top10_markers$gene) + NoLegend()
```



```{r}
DT::datatable(tonsil_markers)
```
## ComplexHeatMap

```{r}
m<-c("SUGCT", "MEF2B", "BCL6", "AICDA","TCL1A", "SOX5","BCL2A1","CD27", "MZB1","XBP1","IRF4","PRDM1","SLAMF7", "MZB1", "FKBP11", "FCRL5", "IGHM","IGHD","FCER2","CXCR4","BANK1","CD83", "PLAC8","FCRL4")

```

```{r}
m<-c("SUGCT", "MEF2B", "BCL6", "AICDA","TCL1A", "SOX5","BCL2A1","CD27", "MZB1","XBP1","IRF4","PRDM1","SLAMF7", "MZB1", "FKBP11", "FCRL5", "IGHM","IGHD","FCER2","CXCR4","BANK1","CD83", "PLAC8","FCRL4")
DefaultAssay(tonsil_wnn_bcell)<-"RNA"

dot <- DotPlot(tonsil_wnn_bcell, features = unique(top5_tonsil_markers$gene),cols = c("lightgrey", "blue"), cluster.idents = T) + theme(axis.text.x = element_text( size = 10, vjust = 0.8, hjust = 0.8)) + ggtitle("Known markers")

dot +
  coord_flip() +
  theme(axis.title = element_blank(), axis.text.y = element_text(size = 5))
```


```{r}
df<- dot$data
head(df)
```
```{r}
### the matrix for the scaled expression 
exp_mat<-df %>% 
  dplyr::select(-pct.exp, -avg.exp) %>%  
  pivot_wider(names_from = id, values_from = avg.exp.scaled) %>% 
  as.data.frame() 
    
row.names(exp_mat) <- exp_mat$features.plot  
exp_mat <- exp_mat[,-1] %>% as.matrix()

head(exp_mat)
```

```{r}
## the matrix for the percentage of cells express a gene

percent_mat<-df %>% 
  dplyr::select(-avg.exp, -avg.exp.scaled) %>%  
  pivot_wider(names_from = id, values_from = pct.exp) %>% 
  as.data.frame() 
    
row.names(percent_mat) <- percent_mat$features.plot  
percent_mat <- percent_mat[,-1] %>% as.matrix()

head(percent_mat)
```

```{r}
## the range is from 0 - 100
range(percent_mat)
```

```{r}
## these two matrix have the same dimension
dim(exp_mat)
```
```{r}
dim(percent_mat)
```


```{r}
## get an idea of the ranges of the matrix
quantile(exp_mat, c(0.1, 0.5, 0.9, 0.99))
```
```{r}
colnames(exp_mat)
```


```{r}

col_fun = circlize::colorRamp2(c(-1, 0, 2), viridis(20)[c(1,10, 20)])


cell_fun = function(j, i, x, y, w, h, fill){
          grid.rect(x = x, y = y, width = w, height = h, 
                    gp = gpar(col = NA, fill = NA))
          grid.circle(x=x,y=y,r= pindex(percent_mat, i, j)/100 * min(unit.c(w, h)),
                      gp = gpar(fill = col_fun(exp_mat[i, j]), col = NA))}
## also do a kmeans clustering for the genes with k = 4
cell_type<-c("NBC","MBC", "GC", "PC")
column_ha<- HeatmapAnnotation(
    cell_type =cell_type,
    col = list(cell_type=setNames(brewer.pal(4, "Paired"), unique(cell_type))
    ),
    na_col = "grey"
)
# anno_simple(1:10, pch = c(1:4, NA, 6:8, NA, 10, 11))

Heatmap(exp_mat,
        heatmap_legend_param=list(title=""),
        column_title = "Top5 DEG", 
        col=col_fun,
        rect_gp = gpar(type = "none"),
        cell_fun = cell_fun,
        row_names_gp = gpar(fontsize = 8),
       #row_km = 6,
        #border = "black",
        show_column_dend = FALSE,
        row_gap = unit(2, "mm"),
       top_annotation = column_ha,
       width = ncol(exp_mat)*unit(15, "mm"), 
        height = nrow(exp_mat)*unit(5, "mm"),
       show_column_names = FALSE)

```
```{r}
p<-Heatmap(exp_mat,
        heatmap_legend_param=list(title=""),
        column_title = "Top5 DEG", 
        col=col_fun,
        rect_gp = gpar(type = "none"),
        cell_fun = cell_fun,
        row_names_gp = gpar(fontsize = 5),
       #row_km = 6,
        #border = "black",
        show_column_dend = FALSE,
        row_gap = unit(2, "mm"),
       top_annotation = column_ha,
       width = ncol(exp_mat)*unit(15, "mm"), 
        height = nrow(exp_mat)*unit(5, "mm"),
       show_column_names = FALSE)

png(paste("/Users/mlromeror/Documents/multiome_tonsil_Lucia/Figures/
", "Dotplot_top5_bcells.png", sep=""),units="in", width=5, height=5,res=300)
p
dev.off()
```
```{r}
calc_ht_size = function(ht, unit = "inch") {
    pdf(NULL)
    ht = draw(ht)
    w = ComplexHeatmap:::width(ht)
    w = convertX(w, unit, valueOnly = TRUE)
    h = ComplexHeatmap:::height(ht)
    h = convertY(h, unit, valueOnly = TRUE)
    dev.off()

    c(w, h)
}
size2 = calc_ht_size(p)
size2
```

```{r}
pdf(paste("/Users/mlromeror/Documents/multiome_tonsil_Lucia/Figures/
", "Dotplot_bcells.pdf", sep=""),width=5,height=6,paper="special")
p
dev.off()
```

```{r}
ha = HeatmapAnnotation(foo = anno_simple(1:10, pch = 1, 
    pt_gp = gpar(col = "red"), pt_size = unit(1:10, "mm")))
draw(ha)

ha
```
```{r}

col_fun = circlize::colorRamp2(c(-1, 0, 2), viridis(20)[c(1,10, 20)])


cell_fun = function(j, i, x, y, w, h, fill){
          grid.rect(x = x, y = y, width = w, height = h, 
                    gp = gpar(col = NA, fill = NA))
          grid.circle(x=x,y=y,r= pindex(percent_mat, i, j)/100 * min(unit.c(w, h)),
                      gp = gpar(fill = col_fun(exp_mat[i, j]), col = NA))}
## also do a kmeans clustering for the genes with k = 4

cluster_anno<-c("NBC","MBC", "GC", "PC")
column_ha<- HeatmapAnnotation(
    cell_type = cell_type,
    col = list(cluster_anno=setNames(brewer.pal(4, "Paired"), unique(cluster_anno))
    ),
    na_col = "grey"
)



Heatmap(exp_mat,
        heatmap_legend_param=list(title=""),
        column_title = "clustered dotplot", 
        col=col_fun,
        rect_gp = gpar(type = "none"),
        cell_fun = cell_fun,
        row_names_gp = gpar(fontsize = 5),
       km = 4,
        #border = "black",
        show_column_dend = FALSE,
       #show_row_dend = FALSE,
        row_gap = unit(2, "mm"),
       top_annotation = column_ha,
       show_column_names = FALSE)
```

```{r}
ha = HeatmapAnnotation(
    Sex = c("male","male","male","male","male"), 
    Age_group = c("kid","kid","young_adult","young_adult","adult"),
    Hospital=c("Clinica Universitaria de Navarra-Pamplona","Clinica Universitaria de Navarra-Pamplona","Clinica Universitaria de Navarra-Pamplona","Clinica Universitaria de Navarra-Pamplona", "H. Cinic-Barcelona"),
    col = list(sex= "blue",
               Age_group = col_fun,
               Hospital = c("Clinica Universitaria de Navarra-Pamplona" = "orange", "H. Cinic-Barcelona" = "red")) )ç
ha

```

```{r}
p <- Heatmap(exp_mat,
        heatmap_legend_param=list(title="expression"),
        column_title = "clustered dotplot", 
        col=col_fun,
        rect_gp = gpar(type = "none"),
        cell_fun = cell_fun,
        row_names_gp = gpar(fontsize = 5),
       # row_km = 6,
  
        show_column_dend = FALSE,
        row_gap = unit(2, "mm"))
#rasterize(p, dpi = 300)
pdf(paste("/Users/mlromeror/Documents/multiome_tonsil_Lucia/Figures/
", "Dotplot_all_data.pdf", sep=""),width=5,height=5,paper="special")
p
dev.off()
```

```{r}

```

```{r}
cluster_anno<-c("NBC","MBC", "GC", "PC")
column_ha<- HeatmapAnnotation(
    cluster_anno = cluster_anno,
    col = list(cluster_anno=setNames(brewer.pal(4, "Paired"), unique(cluster_anno))
    ),
    na_col = "grey"
)
Heatmap(exp_mat, name = "exp", 
        
        row_names_gp = gpar(fontsize = 6),
        show_column_dend = FALSE,
        col=col_fun,
        top_annotation =column_ha)

```
```{r}
p <- Heatmap(exp_mat, name = "exp", 
        
        row_names_gp = gpar(fontsize = 6),
        show_column_dend = FALSE,
        col=col_fun,
        top_annotation =column_ha)
#rasterize(p, dpi = 300)
pdf(paste("/Users/mlromeror/Documents/multiome_tonsil_Lucia/Figures/
", "Heatmap_Bcells_markers2.pdf", sep=""),width=5,height=5,paper="special")
p
dev.off()
```

```{r}

# anno_block_list <- list(subtype =
#                           anno_block(gp = gpar(fill=rep(c("#a6cee3", "#1f78b4","#b2df8a", 
#              "#33a02c"),each=1), col="white"),
#                                      labels = rep(c("NBC","MBC", "GC", "PC"),each=1),
#                                      labels_gp = gpar(col = "black", fontsize=12), which = "row"))

col = c("NBC"="#a6cee3", "MBC"="#1f78b4" ,"GC"= "#b2df8a", "PC"= "#33a02c")

Heatmap(exp_mat, column_km = 4, column_split = sample(c("NBC","MBC", "GC", "PC"), 10, replace = fal)) + 
rowAnnotation(cluster_anno = anno_block(
	graphics = function(index, levels) {
		grid.rect(gp = gpar(fill = col[levels[2]], col = "black")
			gp = gpar(col = col[levels[1]]))
	}
))


```


```{r}
#col = c("foo" = "red", "bar" = "blue")
split = rep(1:4, each = 1)

Heatmap(exp_mat, name="exp",
        
             row_names_gp = gpar(fontsize = 7),
            column_split = split, 
            col=c("lightgrey","blue"),
            show_column_dend = FALSE,
            row_split = 4,
            column_title = "Heatmap B cell markers",
            #column_labels = FALSE,
            top_annotation= HeatmapAnnotation(cluster = anno_block(
    align_to = list(NBC = 1, MBC = 2,GC=3,PC=4),
    panel_fun = function(index, nm) {
    grid.rect(gp = gpar(fill = col[nm]))
      grid.text(nm, 0.5, 0.5,gp = gpar(fontsize = 10,col="white"))
  },
  width = unit(2, "cm"))))
  
 
```
**Save to PDF**

```{r}
p <- Heatmap(exp_mat, name="exp",
             row_names_gp = gpar(fontsize = 7),
            column_split = split, 
            col=col_fun,
            show_column_dend = FALSE,
            row_split = 4,
            column_title = "Heatmap B cell markers",
            #column_labels = FALSE,
            top_annotation= HeatmapAnnotation(cluster = anno_block(
    align_to = list(NBC = 1, MBC = 2,GC=3,PC=4),
    labels_gp = gpar(col = "white", fontsize = 10),
    panel_fun = function(index, nm) {
    grid.rect(gp = gpar(fill = col[nm]))
      grid.text(nm, 0.5, 0.5)
  },
  width = unit(2, "cm"))))
#rasterize(p, dpi = 300)
pdf(paste("/Users/mlromeror/Documents/multiome_tonsil_Lucia/Figures/
", "Heatmap_Bcells_markers1.pdf", sep=""),width=5,height=5,paper="special")
p
dev.off()
```


```{r eval=FALSE, include=FALSE}
Heatmap(matrix(rnorm(100), 10), 
    top_annotation = HeatmapAnnotation(foo = anno_block(gp = gpar(fill = 2:9),
        labels = c("group1", "group2", "group3"), 
        labels_gp = gpar(col = "white", fontsize = 10))),
    column_km = 3,
    left_annotation = rowAnnotation(foo = anno_block(gp = gpar(fill = 2:4),
        labels = c("group1", "group2", "group3"), 
        labels_gp = gpar(col = "white", fontsize = 10))),
    row_km = 3)
```

# Markers exploration

```{r}

function(tonsil_markers,n){
  top_tonsil_markers<-tonsil_markers %>% group_by(cluster) %>% top_n(n = n, wt = avg_log2FC)

}
top7mark_cluster0<-top7_tonsil_markers[["gene"]][1:7]
top7mark_cluster1<-top7_tonsil_markers[["gene"]][8:14]
top7mark_cluster2<-top7_tonsil_markers[["gene"]][15:21]
top7mark_cluster3<-top7_tonsil_markers[["gene"]][22:28]
top7mark_cluster4<-top7_tonsil_markers[["gene"]][29:35]

```


```{r}

DefaultAssay(tonsil_wnn_bcell)<-"RNA"
markers_gg <- function(x){purrr::map(x, function(x) {
  
  p <- FeaturePlot(
    tonsil_wnn_bcell,
    features = x,
    reduction = "wnn.umap",
    pt.size = 0.1
  )
  p
})}
```


## Iñaki markers 

```{r}
m<-c("PRDM1","XBP1","IRF4","MEF2B","BCL6")

DZ<-c("SUGCT", "CXCR4", "AICDA")

LZ<- c("CD83","BCL2A1")

GC<- c("MEF2B", "BCL6","IRF4")

PC<- c("PRDM1","SLAMF7", "MZB1", "FKBP11")
MBC<-c(	"FCRL4", "FCRL5", "PLAC8", "SOX5","CD27", "TNFRSF13B")

NBC<-c("TCL1A","IGHD","IGHM","FCER2")

GCBC<-c("CXCR4", "CD83", "MEF2B", "BCL6")

```

### Dark Zone

```{r}
markers_gg(DZ)

```

### Light Zone

```{r}
markers_gg(LZ)

```

### Germinal Center


```{r}
markers_gg(GC)

```
### MBC

```{r}
markers_gg(MBC)

```

### NBC

```{r}
markers_gg(NBC)

```

### Plasma cell


```{r}
markers_gg(PC)

```

## Pauli markers


```{r}
naive_mem_bcell<-c("BANK1", "FCER2")
cd4_tcell<-c("CD3D", "IL7R")
dz_gc_bcell<-c("MKI67", "TOP2A")
lz_gc_bcell<-c("MARCKSL1", "RGS13", "LMO2", "CCDC88A")	
cytotoxic<-c("GNLY", "NKG7", "GZMK", "CD8A")
MBC<-c(	"FCRL4", "FCRL5", "PLAC8", "SOX5","CD27", "TNFRSF13B")
NBC<-c("TCL1A","IGHD","IGHM","FCER2")
GCBC<-c("CXCR4", "CD83", "MEF2B", "BCL6")
PC<-c("IGHG1", "IGHA1", "JCHAIN", "XBP1")
myeloid<-c("LYZ", "S100A8")	
doublet_proliferative_tcell<-c("MT2A", "CD3D", "TRAC", "PCNA")
PDC<-c("PTCRA", "LILRA4", "IRF7")

```

```{r}
markers_gg(c("IGHM","IL4R"))

```

## Naive and Memory B cell 


```{r}
markers_gg(NBC)
```

```{r}
markers_gg(MBC)
```

### Germinal Center

```{r}
markers_gg(GCBC)
```
### Dark Zone B cell

```{r}

markers_gg(dz_gc_bcell)
```
### Light zone Bcell
```{r}
markers_gg(lz_gc_bcell)
```

### PC

```{r}
markers_gg(PC)
```



 Proliferating cell nuclear antigen (PCNA)
 

```{r}

markers_gg(PDC)

```



## Top 7 of each cluster 

### Cluster 0

FCRL4 is an immunoregulatory receptor that belongs to the Fc receptor-like (FCRL) family. In healthy individuals, FCRL4 is specifically expressed by memory B cells (MBCs) localized in sub-epithelial regions of lymphoid tissues. Expansion of FCRL4+ B cells has been observed in blood and other tissues in various infectious and autoimmune disorders

```{r}
markers_gg(top7mark_cluster0)
```




```{r}
markers_gg(c("MS4A1","NT5E"))
```
MS4A1:NAIVE-MEMORY B CELL
NT5E: NAIVE B CELL 

### Cluster 1: 


```{r}
markers_gg(top7mark_cluster1)
```


### Cluster 2: 


```{r}
markers_gg(top7mark_cluster2)
```



### Cluster 3:

```{r}
markers_gg(top7mark_cluster3)
```



### Cluster 4


```{r}
markers_gg(top7mark_cluster4)
```




## Markers FOXO1, NFKB1

```{r}
 p <- FeaturePlot(
    tonsil_wnn_bcell,
    features = c("NFKB1","FOXO1"),
    reduction = "wnn.umap",
    pt.size = 0.1
  )
p
```


# Bibliography  markers

**MARKERS**

Immature B cells express CD19, CD 20, CD34, CD38, and CD45R,

* **naive B-cells** (identified by high expression of CD79A, CD79B and BLNK).
* **plasma cells** (identified by B-cell and proliferation markers, such as TOP2A or MKI67).
* poor-quality cells (identified by high mitochondrial expression). If a cell has pores in the membrane due to cell lysis, the cytosolic mRNA will leak out of the cell; but if the diameter of mitochondria is higher than the pores, they will get trapped inside the cell.



```{r}
canonical_bcell_markers <-c("CD34", "CD38", "CD19")

monocytes_markers<-c("LYZ","S100A8")

naive_markers<-c("CD79A", "CD79B", "BLNK")

bib_Bcell_markers<-c("CD19","CR2","MS4A1","RALGPS2","CD79A")
bib_Tcell_markers<-c("CD3E","CD4","CD8A","FOXP3","IL17A")

```



## T-cells

T-cell receptor/CD3 complex (TCR/CD3 complex) 

* **T-cells** (identified by high expression of CD3D and CD3E).
* **monocytes** (identified by high expression of LYZ and S100A8).

CD8+ T cell markers:"CD3D", "CD8A"
NK cell markers:"GNLY", "NKG7"

```{r}
markers_gg(bib_Tcell_markers)
```

# Mitocondrial 

```{r}
FeaturePlot(
    tonsil_wnn_bcell,
    features = "percent.mt",
    reduction = "wnn.umap",
    pt.size = 0.1
  )
```

```{r}
VlnPlot(tonsil_wnn_bcell, features = "percent.mt", group.by = "annotation_level_1", pt.size=0,cols = c("#a6cee3", "#1f78b4","#b2df8a", 
             "#33a02c", "#fb9a99","#e31a1c"))+ +NoLegend()+ theme(axis.text.x = element_text(angle = 90))
```

# Ribosomal

```{r}
VlnPlot(tonsil_wnn_bcell, features = "percent_ribo", group.by = "annotation_level_1", pt.size=0, angle)+NoLegend()
```

# Weight of each methodology

```{r}
weight.vp<- VlnPlot(tonsil_wnn_bcell, features = "RNA.weight", group.by = "annotation_level_2",pt.size = 0,cols = cluster_color)
p.weight.vp<- VlnPlot(tonsil_wnn_bcell, features = "peaks.weight", group.by = "annotation_level_2",pt.size = 0,cols = cluster_color)

weight.vp + ggtitle("RNA modality weight")+ theme(axis.text.x = element_text(angle = 0))+ NoLegend()
p.weight.vp+ ggtitle("ATAC modality weight")+ theme(axis.text.x = element_text(angle = 0))+ NoLegend()
```


# Number of cell in each cluster

```{r}
Idents(tonsil_wnn_bcell) <- "annotation_level_1"

cell.num <- table(Idents(tonsil_wnn_bcell))
cell.num
```



# CellCycleScoring

```{r}
DimPlot(tonsil_wnn_bcell,
    reduction = "wnn.umap",
    pt.size = 0.1, label = T, group.by = "Phase")
```

```{r}
DimPlot(tonsil_wnn_bcell,
    reduction = "wnn.umap",
    pt.size = 0.1, label = T)
```

```{r}
FeaturePlot(tonsil_wnn_bcell,
            reduction = "wnn.umap",
            features = "doublet_scores")
```

# Link peaks genes

```{r}
tonsil_wnn_bcell@assays[["peaks"]]<-NULL
```

```{r}
DefaultAssay(tonsil_wnn_bcell) <- "ATAC"
# first compute the GC content for each peak
tonsil_wnn_bcell <- RegionStats(tonsil_wnn_bcell, genome = BSgenome.Hsapiens.UCSC.hg38)

# link peaks to genes
tonsil_wnn_bcell <- LinkPeaks(
  object = tonsil_wnn_bcell,
  peak.assay = "ATAC",
  expression.assay = "RNA",
 # genes.use = c("BCL6","PRDM1"),
  distance = 1e+8
)
```



# Save

```{r}
saveRDS(tonsil_wnn_bcell, path_to_obj_save)

```

# Session info

```{r}
sessionInfo()
```
